Refactored tests, added tests for Home component
