+++ src/tests/testUtils.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import { MemoryRouter, Route } from 'react-router-dom';
+import { render } from '@testing-library/react';
+import Main from '../components/Main';
+
+const setup = (initialPath = '/') => {
+  // access history as described in the docs
+  // https://reactrouter.com/web/guides/testing/checking-location-in-tests
+  let history;
+  render(
+    <MemoryRouter initialEntries={[initialPath]}>
+      <Main />
+      <Route
+        path="*"
+        render={(props) => {
+          history = props.history;
+          return null;
+        }}
+      />
+    </MemoryRouter>,
+  );
+  return { history };
+};
+
+export default setup;
